commit mudança dos tipos os atributos id

+++ System_Almoxarifado/src/senai/almoxarife/empity/SolicitacaoMaterial.java
@@ -3,8 +3,9 @@ package senai.almoxarife.empity;
 public class SolicitacaoMaterial {
 	
 	private int quantidade;
-	private int idSolicitacao;
-	private int idMaterial;
+	private Long idSolicitacao;
+	private Long idMaterial;
+	
 	
 	public int getQuantidade() {
 		return quantidade;
@@ -12,17 +13,19 @@ public class SolicitacaoMaterial {
 	public void setQuantidade(int quantidade) {
 		this.quantidade = quantidade;
 	}
-	public int getIdSolicitacao() {
+	public Long getIdSolicitacao() {
 		return idSolicitacao;
 	}
-	public void setIdSolicitacao(int idSolicitacao) {
+	public void setIdSolicitacao(Long idSolicitacao) {
 		this.idSolicitacao = idSolicitacao;
 	}
-	public int getIdMaterial() {
+	public Long getIdMaterial() {
 		return idMaterial;
 	}
-	public void setIdMaterial(int idMaterial) {
+	public void setIdMaterial(Long idMaterial) {
 		this.idMaterial = idMaterial;
 	}
+	
+	
 
 }
